Move cloud storage cache to file_cache module
